+++ packages/design-system-react-native/src/components/BottomSheet/index.ts
@@ -1,0 +1,6 @@
+export { BottomSheet } from './BottomSheet';
+export type {
+  BottomSheetProps,
+  BottomSheetRef,
+  BottomSheetPostCallback,
+} from './BottomSheet.types';