+++ packages/design-system-react-native/src/components/MainActionButton/index.ts
@@ -1,0 +1,2 @@
+export { MainActionButton } from './MainActionButton';
+export type { MainActionButtonProps } from './MainActionButton.types';