+++ packages/design-system-react-native/src/components/index.ts
@@ -52,6 +52,22 @@ export type {
   BadgeWrapperCustomPosition,
 } from './BadgeWrapper';
 
+export { BannerBase } from './BannerBase';
+export type { BannerBaseProps } from './BannerBase';
+
+export { BottomSheet } from './BottomSheet';
+export type {
+  BottomSheetProps,
+  BottomSheetRef,
+  BottomSheetPostCallback,
+} from './BottomSheet';
+
+export { default as BottomSheetDialog } from './BottomSheetDialog';
+export type {
+  BottomSheetDialogProps,
+  BottomSheetDialogRef,
+} from './BottomSheetDialog';
+
 export { BottomSheetFooter, ButtonsAlignment } from './BottomSheetFooter';
 export type {
   BottomSheetFooterProps,
@@ -79,7 +95,7 @@ export {
   BoxBackgroundColor,
   BoxBorderColor,
 } from './Box';
-export type { BoxProps, BoxSpacing } from './Box';
+export type { BoxProps, BoxSpacing, BoxBorderWidth } from './Box';
 
 export { Card } from './Card';
 export type { CardProps } from './Card';
@@ -93,7 +109,7 @@ export type { ButtonBaseProps } from './ButtonBase';
 export { Button, ButtonSize, ButtonVariant } from './Button';
 export type { ButtonProps } from './Button';
 
-export { ButtonIcon, ButtonIconSize } from './ButtonIcon';
+export { ButtonIcon, ButtonIconSize, ButtonIconVariant } from './ButtonIcon';
 export type { ButtonIconProps } from './ButtonIcon';
 
 export { ButtonSemantic, ButtonSemanticSeverity } from './ButtonSemantic';
@@ -114,6 +130,9 @@ export type { InputProps } from './Input';
 export { Label } from './Label';
 export type { LabelProps } from './Label';
 
+export { ListItem, ListItemVerticalAlignment } from './ListItem';
+export type { ListItemProps } from './ListItem';
+
 export { RadioButton } from './RadioButton';
 export type { RadioButtonProps } from './RadioButton';
 
@@ -123,8 +142,17 @@ export type { JazziconProps } from './temp-components/Jazzicon';
 export { Maskicon } from './temp-components/Maskicon';
 export type { MaskiconProps } from './temp-components/Maskicon';
 
+export { MainActionButton } from './MainActionButton';
+export type { MainActionButtonProps } from './MainActionButton';
+
 export { Skeleton } from './Skeleton';
 export type { SkeletonProps } from './Skeleton';
+
+export { SensitiveText, SensitiveTextLength } from './SensitiveText';
+export type { SensitiveTextProps } from './SensitiveText';
+
+export { TabEmptyState } from './TabEmptyState';
+export type { TabEmptyStateProps } from './TabEmptyState';
 
 export { TextButton, TextButtonSize } from './TextButton';
 export type { TextButtonProps } from './TextButton';
@@ -145,9 +173,16 @@ export type { TextFieldProps } from './TextField';
 export { TextOrChildren } from './temp-components/TextOrChildren';
 export type { TextOrChildrenProps } from './temp-components/TextOrChildren';
 
-export { Toast, ToastVariant, ToastContext } from './Toast';
+export {
+  Toast,
+  ToastVariant,
+  ToastCloseButtonVariant,
+  ToastContext,
+  ToastContextWrapper,
+} from './Toast';
 export type {
   ToastOptions,
+  ToastProps,
   ToastRef,
   ToastContextParams,
   ToastLabelOptions,
@@ -158,3 +193,18 @@ export type {
 
 export { ButtonHero, ButtonHeroSize } from './ButtonHero';
 export type { ButtonHeroProps } from './ButtonHero';
+
+export { ButtonFilter } from './ButtonFilter';
+export type { ButtonFilterProps } from './ButtonFilter';
+
+export { ImageOrSvg } from './temp-components/ImageOrSvg';
+export type {
+  ImageOrSvgProps,
+  ImageOrSvgSrc,
+} from './temp-components/ImageOrSvg';
+
+export { Spinner } from './temp-components/Spinner';
+export type { SpinnerProps } from './temp-components/Spinner';
+
+export { BannerAlert, BannerAlertSeverity } from './BannerAlert';
+export type { BannerAlertProps } from './BannerAlert';